fixes + refactoring
