--- factory/common.go
@@ -1,29 +1,0 @@
-package factory
-
-import (
-	"github.com/ElrondNetwork/elrond-go-storage/common"
-	"github.com/ElrondNetwork/elrond-go-storage/storageUnit"
-)
-
-// GetCacherFromConfig will return the cache config needed for storage unit from a config came from the toml file
-func GetCacherFromConfig(cfg common.CacheConfig) storageUnit.CacheConfig {
-	return storageUnit.CacheConfig{
-		Name:                 cfg.Name,
-		Capacity:             cfg.Capacity,
-		SizePerSender:        cfg.SizePerSender,
-		SizeInBytes:          cfg.SizeInBytes,
-		SizeInBytesPerSender: cfg.SizeInBytesPerSender,
-		Type:                 storageUnit.CacheType(cfg.Type),
-		Shards:               cfg.Shards,
-	}
-}
-
-// GetDBFromConfig will return the db config needed for storage unit from a config came from the toml file
-func GetDBFromConfig(cfg common.DBConfig) storageUnit.DBConfig {
-	return storageUnit.DBConfig{
-		Type:              storageUnit.DBType(cfg.Type),
-		MaxBatchSize:      cfg.MaxBatchSize,
-		BatchDelaySeconds: cfg.BatchDelaySeconds,
-		MaxOpenFiles:      cfg.MaxOpenFiles,
-	}
-}
