--- factory/interface.go
@@ -1,12 +1,0 @@
-package factory
-
-import (
-	"github.com/ElrondNetwork/elrond-go-core/core"
-)
-
-// NodeTypeProviderHandler defines the actions needed for a component that can handle the node type
-type NodeTypeProviderHandler interface {
-	SetType(nodeType core.NodeType)
-	GetType() core.NodeType
-	IsInterfaceNil() bool
-}
