--- factory/persisterFactory.go
@@ -1,57 +1,0 @@
-package factory
-
-import (
-	"errors"
-
-	"github.com/ElrondNetwork/elrond-go-storage"
-	"github.com/ElrondNetwork/elrond-go-storage/common"
-	"github.com/ElrondNetwork/elrond-go-storage/leveldb"
-	"github.com/ElrondNetwork/elrond-go-storage/memorydb"
-	"github.com/ElrondNetwork/elrond-go-storage/storageUnit"
-)
-
-// PersisterFactory is the factory which will handle creating new databases
-type PersisterFactory struct {
-	dbType            string
-	batchDelaySeconds int
-	maxBatchSize      int
-	maxOpenFiles      int
-}
-
-// NewPersisterFactory will return a new instance of a PersisterFactory
-func NewPersisterFactory(config common.DBConfig) *PersisterFactory {
-	return &PersisterFactory{
-		dbType:            config.Type,
-		batchDelaySeconds: config.BatchDelaySeconds,
-		maxBatchSize:      config.MaxBatchSize,
-		maxOpenFiles:      config.MaxOpenFiles,
-	}
-}
-
-// Create will return a new instance of a DB with a given path
-func (pf *PersisterFactory) Create(path string) (elrond_go_storage.Persister, error) {
-	if len(path) == 0 {
-		return nil, errors.New("invalid file path")
-	}
-
-	switch storageUnit.DBType(pf.dbType) {
-	case storageUnit.LvlDB:
-		return leveldb.NewDB(path, pf.batchDelaySeconds, pf.maxBatchSize, pf.maxOpenFiles)
-	case storageUnit.LvlDBSerial:
-		return leveldb.NewSerialDB(path, pf.batchDelaySeconds, pf.maxBatchSize, pf.maxOpenFiles)
-	case storageUnit.MemoryDB:
-		return memorydb.New(), nil
-	default:
-		return nil, elrond_go_storage.ErrNotSupportedDBType
-	}
-}
-
-// CreateDisabled will return a new disabled persister
-func (pf *PersisterFactory) CreateDisabled() elrond_go_storage.Persister {
-	return &disabledPersister{}
-}
-
-// IsInterfaceNil returns true if there is no value under the interface
-func (pf *PersisterFactory) IsInterfaceNil() bool {
-	return pf == nil
-}
