+++ testscommon/marshalizerMock.go
@@ -1,4 +1,4 @@
-package mock
+package testscommon
 
 import (
 	"encoding/json"
