+++ testscommon/sweepHandlerStub.go
@@ -1,4 +1,4 @@
-package mock
+package testscommon
 
 // EvictionHandlerStub -
 type EvictionHandlerStub struct {
