+++ testscommon/trieFactory/serializedStoredData.go
