+++ txcache/txListForSender_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/ElrondNetwork/elrond-go-core/data/transaction"
-	"github.com/ElrondNetwork/elrond-go-storage/mock/txcachemocks"
+	"github.com/ElrondNetwork/elrond-go-storage/testscommon/txcachemocks"
 	"github.com/stretchr/testify/require"
 )
