+++ txcache/wrappedTransaction_test.go
@@ -3,7 +3,7 @@ package txcache
 import (
 	"testing"
 
-	"github.com/ElrondNetwork/elrond-go-storage/mock/txcachemocks"
+	"github.com/ElrondNetwork/elrond-go-storage/testscommon/txcachemocks"
 	"github.com/stretchr/testify/require"
 )
